rpc: commands: test file sizes larger than `UINT16_MAX`
Extend tests that use `RPC_DATA` packets to validate they work with
data sizes over `UINT16_MAX`.

+++ subsys/rpc/commands/data_receiver.c
@@ -26,7 +26,7 @@ struct net_buf *rpc_command_data_receiver(struct net_buf *request)
 	enum epacket_auth auth;
 	uint32_t data_offset;
 	uint32_t received = 0;
-	uint16_t expected_offset = 0;
+	uint32_t expected_offset = 0;
 	uint8_t ack_period;
 	size_t var_len;
 	int rc = 0;
